Merged PR 419044: edit pass: compatibility
Please review the copy edit to this article, note the comments, and let me know if I’ve changed the technical meaning anywhere. This is for Azure DevOps work item 1398491.
Comments:
There are instances of the word "should" that we typically try to replace, but I didn't. They seemed necessary, but I wanted to call them to your attention. Line 30 guarantees a level of support.

Related work items: #1398491